remove pdf/confest.py and merge with tests/conftest.py
