+++ flaskapp/app/tests/conftest.py
@@ -1,6 +1,6 @@
 from gql import Client, gql
 from gql.transport.aiohttp import AIOHTTPTransport
-from datetime import datetime
+from datetime import datetime, timezone, timedelta
 from base64 import b64decode
 from app.env import DatabaseSettings
 from sqlalchemy import create_engine
@@ -16,6 +16,7 @@ def client():
     transport = AIOHTTPTransport(url=GRAPHQL_MUTATION_QUERY_URL)
     # Create a GraphQL client using the defined transport
     return Client(transport=transport, fetch_schema_from_transport=True)
+
 
 @pytest.fixture
 def auth_client(client):
@@ -95,3 +96,35 @@ def dbsession(engine, tables):
     transaction.rollback()
     # put back the connection to the connection pool
     connection.close()
+
+
+@pytest.fixture
+def lenght_test():
+    """generate a string with data with charactes to test lenght"""
+    lenght_test = ''
+    for x in range(0, 1100):
+        lenght_test += str(x)
+    return lenght_test
+
+@pytest.fixture
+def datetime_to_use():
+    """get current datetime to test"""
+    return datetime.now().strftime('%d/%m/%Y')
+
+@pytest.fixture
+def datetime_with_timezone_to_use():
+    """get current datetime to test"""
+    tmz = timezone(offset=timedelta(hours=-3))
+    return datetime.now(tz=tmz).strftime('%d/%m/%Y')
+
+@pytest.fixture
+def document_datetime_to_use():
+    """get current datetime with hours and minutes to test"""
+    return datetime.now().strftime('%d/%m/%Y %H:%M')
+
+@pytest.fixture
+def client():
+    # Select your transport with ag graphql url endpoint
+    transport = AIOHTTPTransport(url=GRAPHQL_MUTATION_QUERY_URL)
+    # Create a GraphQL client using the defined transport
+    return Client(transport=transport, fetch_schema_from_transport=True)
